requirements.txt
Flask
pandas
opencv-python
cmake
face_recognition
numpy
mysql-connector-python
twilio